requirements.txt
adbutils>=2.6.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pywebio>=1.8.0
pyyaml>=6.0
